requirements.txt
notion
boardgamegeek2
google-api-python-client
pandas
numpy
requests
google-auth-oauthlib
google-auth-httplib2
loguru
tqdm
